Move enum class
