+++ src/com/app/model/Student.java
@@ -1,5 +1,7 @@
 package com.app.model;
 
+import com.app.controllers.SchoolSubject;
+
 import java.util.ArrayList;
 import java.util.HashMap;
